requirements.txt
flake8
intake-dcat
intake-geopandas
papermill
intake-civis
pandas
boto3

